Fix or hide all remaining warnings.
Disable warning 4267 in CMake (so it affects jpeg & png libs too).
Warning 4267 is: conversion from 'size_t' to 'int', possible loss of data.

Disable (in doomclassic only) warning about unary minus on unsigned types, which was a deliberate hack.

Fix some warnings about conversions from size_t to int that it makes sense to fix.
Fix warnings about conversions from size_t to int in my own code for uncensoring Doom classic.

Fix all the other warnings.

+++ neo/aas/AASFile_local.h
@@ -66,7 +66,7 @@ public:
 	bool						Load( const idStr& fileName, unsigned int mapFileCRC );
 	bool						Write( const idStr& fileName, unsigned int mapFileCRC );
 	
-	int							MemorySize() const;
+	size_t						MemorySize() const;
 	void						ReportRoutingEfficiency() const;
 	void						Optimize();
 	void						LinkReversedReachability();
